Inline steam redirect link
The `/api/login/steam` route is now only used in development. In
production, the route is no longer needed as the steam link is embedded
directly in the frontend.

What originally prompted this change is `/api/login/steam` could be
prerendered at build time. On cloudflare this caused the redirect to be
served as an empty file without any headers.

`EXTERNAL_ADDRESS` is has been moved from a deployment environment to a
build environment.
